pyproject.toml
[tool.poetry]
name = "my-apis"
version = "0.1.0"
description = "This project consist all the hosted apis I develop"
authors = ["Aman Rawat <kirmolaa@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
djangorestframework = "^3.15.1"
django-filter = "^24.2"
markdown = "^3.6"
faker = "^25.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
